vuteco/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vuteco"
version = "0.1.0"
requires-python = "~=3.10.0"
dependencies = [
    "accelerate~=1.6.0",
    "aiohappyeyeballs~=2.6.1",
    "aiohttp~=3.11.16",
    "aiosignal~=1.3.2",
    "async-timeout~=5.0.1",
    "attrs~=25.3.0",
    "bitsandbytes~=0.45.5",
    "certifi~=2025.1.31",
    "charset-normalizer~=3.4.1",
    "click~=8.1.8",
    "colorama~=0.4.6",
    "contourpy~=1.3.0",
    "cut-cross-entropy @ git+https://github.com/apple/ml-cross-entropy.git@058369b06bc0c465fbf0ac372c3a8e18d62df52f",
    "cwe2~=3.0.0",
    "cycler~=0.12.1",
    "datasets~=3.5.0",
    "dill~=0.3.8",
    "docker~=7.1.0",
    "docstring_parser~=0.16",
    "eval_type_backport~=0.2.2",
    "evaluate~=0.4.3",
    "filelock~=3.18.0",
    "fonttools~=4.57.0",
    "frozenlist~=1.5.0",
    "fsspec~=2024.12.0",
    "gitdb~=4.0.12",
    "GitPython~=3.1.44",
    "halo~=0.0.31",
    "hf_transfer~=0.1.9",
    "huggingface-hub~=0.30.2",
    "idna~=3.10",
    "imbalanced-learn~=0.12.4",
    "importlib_metadata~=8.6.1",
    "importlib_resources~=6.5.2",
    "javalang~=0.13.0",
    "jellyfish~=1.2.0",
    "Jinja2~=3.1.6",
    "joblib~=1.4.2",
    "kiwisolver~=1.4.7",
    "log-symbols~=0.0.14",
    "markdown-it-py~=3.0.0",
    "MarkupSafe~=3.0.2",
    "matplotlib~=3.9.4",
    "matplotlib-venn~=1.1.2",
    "mdurl~=0.1.2",
    "mpmath~=1.3.0",
    "multidict~=6.2.0",
    "multiprocess~=0.70.16",
    "networkx~=3.2.1",
    "nltk~=3.9.1",
    "numpy~=2.0.2",
    "nvidia-cublas-cu12~=12.4.5.8",
    "nvidia-cuda-cupti-cu12~=12.4.127",
    "nvidia-cuda-nvrtc-cu12~=12.4.127",
    "nvidia-cuda-runtime-cu12~=12.4.127",
    "nvidia-cudnn-cu12~=9.1.0.70",
    "nvidia-cufft-cu12~=11.2.1.3",
    "nvidia-curand-cu12~=10.3.5.147",
    "nvidia-cusolver-cu12~=11.6.1.9",
    "nvidia-cusparse-cu12~=12.3.1.170",
    "nvidia-cusparselt-cu12~=0.6.2",
    "nvidia-nccl-cu12~=2.21.5",
    "nvidia-nvjitlink-cu12~=12.4.127",
    "nvidia-nvtx-cu12~=12.4.127",
    "packaging~=24.2",
    "pandas~=2.2.3",
    "peft~=0.15.1",
    "pillow~=11.1.0",
    "propcache~=0.3.1",
    "protobuf~=3.20.3",
    "psutil~=7.0.0",
    "pyarrow~=19.0.1",
    "pycvesearch~=1.2",
    "Pygments~=2.19.1",
    "pyparsing~=3.2.3",
    "python-dateutil~=2.9.0.post0",
    "pytz~=2025.2",
    "PyYAML~=6.0.2",
    "regex~=2024.11.6",
    "requests~=2.32.3",
    "rich~=14.0.0",
    "safetensors~=0.5.3",
    "scikit-learn~=1.6.1",
    "scipy~=1.13.1",
    "segtok~=1.5.11",
    "sentencepiece~=0.2.0",
    "shtab~=1.7.1",
    "six~=1.17.0",
    "smmap~=5.0.2",
    "spinners~=0.0.24",
    "sympy~=1.13.1",
    "tabulate~=0.9.0",
    "termcolor~=3.0.1",
    "threadpoolctl~=3.6.0",
    "tokenizers~=0.21.1",
    "torch~=2.6.0",
    "tqdm~=4.67.1",
    "transformers~=4.51.1",
    "triton~=3.2.0",
    "trl~=0.15.2",
    "typeguard~=4.4.2",
    "typing_extensions~=4.13.1",
    "tyro~=0.9.18",
    "tzdata~=2025.2",
    "unsloth @ git+https://github.com/unslothai/unsloth.git@73d6fb2868cbaf08e57f152425421a2db2217b1b",
    "unsloth_zoo @ git+https://github.com/unslothai/unsloth-zoo.git@4a66f8b08952fc148f5c74cd15aec52cb0113e2d",
    "urllib3~=2.3.0",
    "xformers~=0.0.29.post3",
    "xxhash~=3.5.0",
    "yake~=0.4.8",
    "yarl~=1.19.0",
    "zipp~=3.21.0",
]

[project.scripts]
vuteco = "main.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
